map overlays request
