+++ packages/tupaia-web/src/types/types.d.ts
@@ -2,18 +2,7 @@
  * Tupaia
  *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
-import {
-  LandingPage,
-  Project,
-  Country,
-  MapOverlay,
-  TupaiaWebDashboardsRequest,
-  TupaiaWebProjectRequest,
-  TupaiaWebEntityRequest,
-} from '@tupaia/types';
-import { GRANULARITY_CONFIG } from '@tupaia/utils';
-import { ActivePolygonProps } from '@tupaia/ui-map-components';
-import { Position } from 'geojson';
+import { LandingPage, Project, Country, TupaiaWebProjectRequest } from '@tupaia/types';
 import { KeysToCamelCase } from './helpers';
 
 export type SingleProject = TupaiaWebProjectRequest.ResBody & {
@@ -36,32 +25,3 @@ export type CountryAccessListItem = Country & {
   hasAccess: boolean;
   accessRequests: string[];
 };
-
-export type Dashboard = TupaiaWebDashboardsRequest.ResBody[0];
-
-export type SingleMapOverlayItem = KeysToCamelCase<
-  Pick<MapOverlay, 'code', 'name', 'legacy', 'report_code'>
-> & {
-  measureLevel?: string;
-  displayType: string;
-  periodGranularity?: keyof typeof GRANULARITY_CONFIG;
-  startDate?: string;
-  endDate?: string;
-};
-
-export type MapOverlayGroup = {
-  name: MapOverlay['name'];
-  children: SingleMapOverlayItem[] | MapOverlayGroup[];
-};
-
-// re-type the coordinates to be what the ui-map-components expect, because in the types package they are any | null
-export type Entity = KeysToCamelCase<Omit<TupaiaWebEntityRequest.ResBody, 'region' | 'bounds'>> & {
-  region?: ActivePolygonProps['coordinates'];
-  bounds?: Position[];
-  parentCode: Entity['code'];
-  childCodes: Entity['code'][];
-  photoUrl?: string;
-  children?: Entity[];
-};
-
-export type EntityCode = Entity['code'];
